Use absolute symlinks when testing on windows
Relative symlinks on windows confuse filepath.Walk. This is Go issue 17540,
https://github.com/golang/go/issues/17540. While waiting for that to get fixed,
just use absolute symlinks in windows tests.
