+++ filesystem_test.go
@@ -28,32 +28,6 @@ type filesystemState struct {
 	links []fsLink
 }
 
-// setup inflates fs onto the actual host file system
-func (fs filesystemState) setup(t *testing.T) {
-	for _, dir := range fs.dirs {
-		p := dir.prepend(fs.root)
-		if err := os.MkdirAll(p.String(), 0777); err != nil {
-			t.Fatalf("os.MkdirAll(%q, 0777) err=%q", p, 0777)
-		}
-	}
-	for _, file := range fs.files {
-		p := file.prepend(fs.root)
-		f, err := os.Create(p.String())
-		if err != nil {
-			t.Fatalf("os.Create(%q) err=%q", p, err)
-		}
-		if err := f.Close(); err != nil {
-			t.Fatalf("file %q Close() err=%q", p, err)
-		}
-	}
-	for _, link := range fs.links {
-		p := link.path.prepend(fs.root)
-		if err := os.Symlink(link.to, p.String()); err != nil {
-			t.Fatalf("os.Symlink(%q, %q) err=%q", link.to, p, err)
-		}
-	}
-}
-
 // assert makes sure that the fs state matches the state of the actual host
 // file system
 func (fs filesystemState) assert(t *testing.T) {
@@ -71,7 +45,7 @@ func (fs filesystemState) assert(t *testing.T) {
 		linkMap[l.path.prepend(fs.root).String()] = struct{}{}
 	}
 
-	filepath.Walk(fs.root, func(path string, info os.FileInfo, err error) error {
+	err := filepath.Walk(fs.root, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
 			t.Errorf("filepath.Walk path=%q  err=%q", path, err)
 			return err
@@ -111,6 +85,10 @@ func (fs filesystemState) assert(t *testing.T) {
 		}
 		return nil
 	})
+
+	if err != nil {
+		t.Errorf("filesystem.Walk err=%q", err)
+	}
 
 	for d := range dirMap {
 		t.Errorf("could not find expected directory %q", d)
